Se agrega el arbol de archivos
